+++ src/wss/private/mod.rs
@@ -1,5 +1,0 @@
-//! Messages for private websocket channels
-pub mod messages;
-pub mod open_orders_messages;
-pub mod own_trades_messages;
-pub mod trading_messages;