+++ src/wss/public/mod.rs
@@ -1,3 +1,0 @@
-//! Messages for public websocket channels
-pub mod messages;
-pub mod orderbooks;