+++ tests/resources/mod.rs
@@ -1,7 +1,5 @@
 pub mod kraken_responses;
 pub mod test_auth;
 pub mod test_client;
-pub mod test_logging;
 pub mod test_macros;
 pub mod test_serde;
-pub mod wss;